fix(paper): FR-2 — PaperExitEvent monetary fields Decimal-as-TEXT
Why: float/REAL columns on paper_exit_events violated the Decimal-as-TEXT
invariant (REVIEW.md §5) and blocked SIG-2's end-to-end Decimal conversion.

What:
- src/paper/models.py: ltp, mid, bid, ask, entry_price, threshold_value → Decimal|None
- src/paper/store.py: schema columns REAL→TEXT; create_exit_event params Decimal;
  serialize with str(); add _parse_exit_event_row helper; wire into get_open_exit_events
- src/strategy/reentry_mixin.py: entry_price=0.0 → Decimal('0'); import Decimal
- scripts/strategies/three_track/paper_3track_snapshot.py: drop float() wrappers
  on opt_leg.ltp/bid/ask; threshold_value via Decimal(str(...)); threshold=Decimal('5')
- scripts/dev/migrate_exit_events_decimal.py: rename-recreate migration for live DB
- tests/unit/paper/test_paper_store_exit_events.py: update all float assertions;
  add test_monetary_fields_stored_as_text_in_db, test_decimal_roundtrip_preserves_exact_value,
  test_none_monetary_fields_stored_as_null
Ref: REVIEW.md §5 Decimal-as-TEXT invariant

+++ src/paper/models.py
@@ -220,14 +220,14 @@ class PaperExitEvent(BaseModel):
     detected_by: Literal["EOD", "INTRADAY", "MANUAL"]
     exit_signal: ExitSignal
     severity: Literal["INFO", "WARNING", "ACTION"]
-    ltp: float | None = None
-    mid: float | None = None
-    bid: float | None = None
-    ask: float | None = None
+    ltp: Decimal | None = None
+    mid: Decimal | None = None
+    bid: Decimal | None = None
+    ask: Decimal | None = None
     delta: float | None = None
     dte: int | None = None
-    entry_price: float
-    threshold_value: float | None = None
+    entry_price: Decimal
+    threshold_value: Decimal | None = None
     delta_stop_would_fire: Literal[0, 1] | None = None
     premium_stop_would_fire: Literal[0, 1] | None = None
     actual_rule_used: str | None = None
